requirements.txt
celery==4.1.0
kombu==4.1.0
pymssql==2.1.4
SQLAlchemy==1.2.16
Flask==1.0.2
Flask-Admin==1.5.0
Flask-Login==0.4.0
Flask-Restless==0.17.0
Flask-SQLAlchemy==2.3.1
Flask-Script==2.0.6
Flask-WTF==0.14.2
Flask-Mail==0.9.1
Flask-marshmallow==0.9.0
Flask-HTTPAuth==3.2.3
flask_cors==3.0.3
marshmallow-sqlalchemy==0.15.0
boto3==1.4.7
psycopg2==2.7.6.1
cryptography==2.0.3
requests==2.21.0
passlib==1.7.1
itsdangerous==0.24
zope.interface==4.4.2
thrift==0.10.0
pyhive==0.5.0
ha_jaydebeapi==2.1.7
ldap3==2.4
jprops==2.0.2
PyJWT==1.5.3
sanic==18.12.0
flasgger==0.8.3
datadog==0.26.0
gunicorn==19.9.0
thrift_sasl==0.3.0
git+https://github.com/mher/flower
urllib3==1.24.1
great_expectations==0.4.5
algoliasearch==1.20.0
